refactor: extract ErrorPageFactory from WebPageFactory
Moves 2 error page methods (~42 lines) into a dedicated ErrorPageFactory
with zero constructor dependencies. WebPageFactory delegates to it.
