+++ platform-web/src/main/kotlin/io/github/rygel/outerstellar/platform/web/WebPageFactory.kt
@@ -153,49 +153,9 @@ open class WebPageFactory(
     fun buildResetPasswordPage(ctx: WebContext, token: String): Page<ResetPasswordPage> =
         authPageFactory.buildResetPasswordPage(ctx, token)
 
-    fun buildErrorPage(ctx: WebContext, kind: String): Page<ErrorPage> {
-        val i18n = ctx.i18n
-        val shell = ctx.shell(i18n.translate("web.nav.errors"), "/errors")
-        val normalizedKind = if (kind == "server-error") "server-error" else "not-found"
-        val statusCode =
-            if (normalizedKind == "server-error") {
-                HTTP_STATUS_SERVER_ERROR
-            } else {
-                HTTP_STATUS_NOT_FOUND
-            }
+    fun buildErrorPage(ctx: WebContext, kind: String): Page<ErrorPage> = errorPageFactory.buildErrorPage(ctx, kind)
 
-        return Page(
-            shell = shell,
-            data =
-                ErrorPage(
-                    statusCode = statusCode,
-                    heading = i18n.translate("web.error.$normalizedKind.title"),
-                    message = i18n.translate("web.error.$normalizedKind.message"),
-                    primaryActionLabel = i18n.translate("web.error.primary"),
-                    primaryActionUrl = ctx.url("/"),
-                    secondaryActionLabel = i18n.translate("web.error.secondary"),
-                    secondaryActionUrl = ctx.url("/auth"),
-                    helpButtonLabel = i18n.translate("web.error.help"),
-                    helpUrl = ctx.url("/errors/components/help/$normalizedKind"),
-                    errorLabel = i18n.translate("web.error.label"),
-                ),
-        )
-    }
-
-    fun buildErrorHelp(ctx: WebContext, kind: String): ErrorHelpFragment {
-        val i18n = ctx.i18n
-        val normalizedKind = if (kind == "server-error") "server-error" else "not-found"
-
-        return ErrorHelpFragment(
-            title = i18n.translate("web.error.$normalizedKind.help.title"),
-            items =
-                listOf(
-                    i18n.translate("web.error.$normalizedKind.help.item1"),
-                    i18n.translate("web.error.$normalizedKind.help.item2"),
-                    i18n.translate("web.error.$normalizedKind.help.item3"),
-                ),
-        )
-    }
+    fun buildErrorHelp(ctx: WebContext, kind: String): ErrorHelpFragment = errorPageFactory.buildErrorHelp(ctx, kind)
 
     @Suppress("LongParameterList")
     fun buildMessageList(
@@ -363,9 +323,10 @@ open class WebPageFactory(
         )
     }
 
-    // Delegated admin methods - kept for backward compatibility
+    // Delegated admin/error/auth methods - kept for backward compatibility
     private val adminPageFactory by lazy { AdminPageFactory(securityService, notificationService) }
     private val authPageFactory by lazy { AuthPageFactory() }
+    private val errorPageFactory by lazy { ErrorPageFactory() }
 
     fun buildUserAdminPage(ctx: WebContext, limit: Int = 20, offset: Int = 0): Page<UserAdminPage> =
         adminPageFactory.buildUserAdminPage(ctx, limit, offset)
